Reorganize the quantization directory
