+++ rs/index/src/hnsw/reader.rs
@@ -100,8 +100,8 @@ mod tests {
     use std::fs;
 
     use quantization::no_op::{NoQuantizer, NoQuantizerWriter};
-    use quantization::pq::{ProductQuantizer, ProductQuantizerConfig, ProductQuantizerWriter};
-    use quantization::pq_builder::{ProductQuantizerBuilder, ProductQuantizerBuilderConfig};
+    use quantization::pq::pq::{ProductQuantizer, ProductQuantizerConfig, ProductQuantizerWriter};
+    use quantization::pq::pq_builder::{ProductQuantizerBuilder, ProductQuantizerBuilderConfig};
     use utils::test_utils::generate_random_vector;
 
     use super::*;
