+++ rs/quantization/src/lib.rs
@@ -1,9 +1,7 @@
 #![feature(portable_simd)]
 pub mod no_op;
 pub mod pq;
-pub mod pq_builder;
 pub mod quantization;
 pub mod rabitq;
-pub mod rabitq_builder;
 pub mod reader;
 pub mod typing;
