+++ rs/quantization/src/pq/mod.rs
@@ -1,0 +1,2 @@
+pub mod pq;
+pub mod pq_builder;
