+++ rs/quantization/src/rabitq/mod.rs
@@ -1,0 +1,2 @@
+pub mod rabitq;
+pub mod rabitq_builder;
